Moved Temp source to ADCs
Big changes

+++ TMS_MAIN/Core/Inc/thermistor.h
@@ -27,7 +27,7 @@ int8_t thermistor_raw_to_c(uint16_t raw);
  * and valid[] are CAN_TMS_THERMS_PER_MODULE long; invalid packs are excluded
  * from low/high/avg. therm_count is always 6 per spec. high_id/low_id index the
  * hottest/coldest pack (0..5). */
-void thermistor_build_module_data(const uint16_t *raw, const uint8_t *valid,
+void thermistor_build_module_data(const uint32_t *raw, const uint8_t *valid,
                                   can_tms_module_data_t *out);
 
 #endif /* INC_THERMISTOR_H_ */
